requirements.txt
anyio==4.2.0
attrs==23.2.0
audioread==3.0.1
auto-subtitle==1.0
Babel==2.14.0
bibtexparser==2.0.0b6
bing-image-downloader==1.1.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.1
colorama==0.4.6
colorlog==6.8.0
contourpy==1.2.0
csvw==3.3.0
cycler==0.12.1
decorator==4.4.2
dlinfo==1.2.1
einops==0.7.0
einops-exts==0.0.4
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.1
flask-ngrok==0.0.25
fonttools==4.47.2
fsspec==2023.12.2
future==0.18.3
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
monotonic_align @ git+https://github.com/resemble-ai/monotonic_align.git@78b985be210a03d08bc3acc01c4df0442105366f
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
munch==4.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==23.2
phonemizer==3.2.1
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
proglog==0.1.10
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
pylatexenc==2.10
pyngrok==7.0.5
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.17.1
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.12.0
segments==2.2.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
websockets==11.0.3
Werkzeug==3.0.1
